[tools] perf: Improve robustness of topology parsing code
Message-id: <1389109877-9395-140-git-send-email-jolsa@redhat.com>
Patchwork-id: 73440
O-Subject: [PATCH RHEL7.0 BZ1036665 139/226] perf tools: Improve robustness of topology parsing code
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit c5885749e4ebe568cca969d43488a233e69e6454
Author: Stephane Eranian <eranian@google.com>
Date: Wed Aug 14 12:04:26 2013 +0200

description
===========
This patch improves the robustness of the build_cpu_topo() routine by
allowing either the CPU parsing or the thread parsing to fail and yet
get perf to produce some topology data which could be useful for the
analysis.

Without this patch, if the cpu parsing fails, the thread parsing is not
attempted vice-versa.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>