[tools] perf: change machine__findnew_thread() to set thread pid
Message-id: <1389109877-9395-163-git-send-email-jolsa@redhat.com>
Patchwork-id: 73463
O-Subject: [PATCH RHEL7.0 BZ1036665 162/226] perf tools: change machine__findnew_thread() to set thread pid
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit 314add6b1f045b59ca39683bd0cbc5310cd203f2
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Aug 27 11:23:03 2013 +0300

description
===========
Add a new parameter for 'pid' to machine__findnew_thread().
Change callers to pass 'pid' when it is known.

Note that callers sometimes want to find the main thread
which has the memory maps.  The main thread has tid == pid
so the usage in that case is:

	machine__findnew_thread(machine, pid, pid)

whereas the usage to find the specific thread is:

	machine__findnew_thread(machine, pid, tid)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>