[tools] perf/evlist: Move perf_evlist__config() to a new source file
Message-id: <1389109877-9395-167-git-send-email-jolsa@redhat.com>
Patchwork-id: 73468
O-Subject: [PATCH RHEL7.0 BZ1036665 166/226] perf evlist: Move perf_evlist__config() to a new source file
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit faf967068e8c4d8df52f01f9361241101b3065a0
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Aug 27 11:23:08 2013 +0300

description
===========
perf_evlist__config() must be moved to a separate source file to avoid
Python link errors when adding support for PERF_SAMPLE_IDENTIFIER.

It is appropriate to do this because perf_evlist__config() is a helper
function for event recording.  It is used by tools to apply recording
options to perf_evlist.  It is not used by the Python API.
---
 create mode 100644 tools/perf/util/record.c

Signed-off-by: Jarod Wilson <jarod@redhat.com>