[tools] perf/trace: Handle perf.data files with no tracepoints
Message-id: <1389109877-9395-192-git-send-email-jolsa@redhat.com>
Patchwork-id: 73491
O-Subject: [PATCH RHEL7.0 BZ1036665 191/226] perf trace: Handle perf.data files with no tracepoints
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit ccf53eac2097fb47bc40875ffb22c2d10fa8c46c
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Fri Sep 6 15:19:01 2013 -0300

description
===========
Before:

  perf trace -i perf.data
  Segmentation fault (core dumped)
  #

After:

 # perf trace -i perf.data
 Data file does not have raw_syscalls:sys_enter events
 #

When there are no tracepoints in a perf.data file the struct pevent
that contains the list of tracepoints that will be used to lookup the
tracepoint id by name will not be populated, causing a NULL deref.

And we don't need to do all that dance to look at pevents for an entry
with a slighly different name to then lookup the tracepoint by its id on
the evlist, just use the perf_evlist__find_tracepoint_by_name() routine,
that will find the tracepoint, if present.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>