[tools] perf/probe: Fix to initialize fname always before use it
Message-id: <1389109877-9395-218-git-send-email-jolsa@redhat.com>
Patchwork-id: 73520
O-Subject: [PATCH RHEL7.0 BZ1036665 217/226] perf probe: Fix to initialize fname always before use it
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit 1b286bdd5b7684c681b63d5e75cada037064c315
Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Date: Fri Oct 11 12:23:17 2013 +0000

description
===========
Fix perf probe --list to initialize fname local var always before
use it. This may cause a SEGV if there is a probe which is in
the function body but not in any inline function.

Problem introduced in:

	commit e08cfd4bda76
	Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
	Date:   Mon Sep 30 18:21:44 2013 +0900

	    perf probe: Fix to find line information for probe list
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>