[tools] perf: Disable PERF_RECORD_MMAP2 support
Message-id: <1389109877-9395-220-git-send-email-jolsa@redhat.com>
Patchwork-id: 73514
O-Subject: [PATCH RHEL7.0 BZ1036665 219/226] perf: Disable PERF_RECORD_MMAP2 support
Bugzilla: 1036665
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1036665
https://bugzilla.redhat.com/show_bug.cgi?id=1036665

upstream
========
commit 3090ffb5a2515990182f3f55b0688a7817325488
Author: Stephane Eranian <eranian@google.com>
Date: Thu Oct 17 19:32:15 2013 +0200

description
===========
For now, we disable the extended MMAP record support (MMAP2).

We have identified cases where it would not report the correct mapping
information, clone(VM_CLONE) but with separate pids.  We will revisit
the support once we find a solution for this case.

The patch changes the kernel to return EINVAL if attr->mmap2 is set. The
patch also modifies the perf tool to use regular PERF_RECORD_MMAP for
synthetic events and it also prevents the tool from requesting
attr->mmap2 mode because the kernel would reject it.

The support will be revisited once the kenrel interface is updated.

In V2, we reduce the patch to the strict minimum.

In V3, we avoid calling perf_event_open() with mmap2 set because we know
it will fail and require fallback retry.
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>