[ethernet] i40e: Do not directly increment Tx next_to_use
Message-id: <1389262521-17570-10-git-send-email-sassmann@redhat.com>
Patchwork-id: 73729
O-Subject: [RHEL7.0 PATCH v2 009/152] i40e: Do not directly increment Tx next_to_use
Bugzilla: 1011561
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Alexander Duyck <alexander.h.duyck@intel.com>

Avoid directly incrementing next_to_use for multiple reasons.  The main
reason being that if we directly increment it then it can attain a state
where it is equal to the ring count.  Technically this is a state it
should not be able to reach but the way this is written it now can.

This patch pulls the value off into a register and then increments it
and writes back either the value or 0 depending on if the value is equal
to the ring count.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Tested-by: Kavindya Deegala <kavindya.s.deegala@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Stefan Assmann <sassmann@redhat.com>
(cherry picked from commit fc4ac67bc98f1d54ec0ef8bcddf54892ad78e1f9)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>