[kernel] kdump: Increase x86 crashkernel=auto base reservation from 128M to 160M
Message-id: <1389355279-19477-1-git-send-email-bhe@redhat.com>
Patchwork-id: 73922
O-Subject: [RHEL7.0 Patch] Increase crashkernel=auto base reservation from 128M to 160M
Bugzilla: 1044832
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>
RH-Acked-by: Dave Young <dyoung@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Resolves: bz1044832
https://bugzilla.redhat.com/show_bug.cgi?id=1044832

In RHEL7, how to generate initramfs and what to pack into it have
fundamentally changed compared to RHEL6. However crashkernel=auto still
follows RHEL6, which uses 128M as the base reservation memory. The result is
we're starting to face the problem that memory is running low in 2nd kernel.
Especially in the case of nfs dump, we've seen OOM from time to time.

In RHEL6, if we unpack kdump initrd for nfs dump, there's totally 11 MB.
Unfortunately the number increased to 43M in RHEL7. The is because: 1. We are
using bash over busybox, systemd over init script and a lot of other new
binaries.  2. Most of binaries and libraries are bloating and their size is
much larger.  3. They're depending on more libraries.

Another stress from user space is binaries and libraries are consuming much
more memory than they do in RHEL6. For instance, dhclient in RHEL7 is cosuming
12MB, however it's only several hundreds of KB when we're using busybox udhcpc
in RHEL6. systemd-journald is consuming about 6MB however back in RHEL6, we
don't use this at all. This kind of difference is everywhere between RHEL7 and
RHEL6.

After analysis, we decide to increase crashkernel=auto base reservation memory
by 32M, that is from 128M to 160M for RHEL7. For the time being, it's efficient.
And we can also take the advantage of more free memory to improve dumping
performance in makedumpfile.

Brew build:
https://brewweb.devel.redhat.com/taskinfo?taskID=6839653

Test:
Passed on dell-pe860-01.rhts.eng.bos.redhat.com.

Signed-off-by: Jarod Wilson <jarod@redhat.com>