[kernel] audit: remove CONFIG_AUDIT_LOGINUID_IMMUTABLE
Message-id: <b1d7a4cea2c1253b590abe470351ff17e3b79edb.1389385575.git.rgb@redhat.com>
Patchwork-id: 74000
O-Subject: [RHEL7 PATCH 3/9] audit: remove CONFIG_AUDIT_LOGINUID_IMMUTABLE
Bugzilla: 1045670
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Eric Paris <eparis@redhat.com>

From: Eric Paris <eparis@redhat.com>

After trying to use this feature in Fedora we found the hard coding
policy like this into the kernel was a bad idea.  Surprise surprise.
We ran into these problems because it was impossible to launch a
container as a logged in user and run a login daemon inside that container.
This reverts back to the old behavior before this option was added.  The
option will be re-added in a userspace selectable manor such that
userspace can choose when it is and when it is not appropriate.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>