[kernel] audit: audit feature to only allow unsetting the loginuid
Message-id: <9d834fc4ba2d8d1073b24cf0c772a4a1bda71140.1389385575.git.rgb@redhat.com>
Patchwork-id: 74003
O-Subject: [RHEL7 PATCH 5/9] audit: audit feature to only allow unsetting the loginuid
Bugzilla: 1045670
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Eric Paris <eparis@redhat.com>

From: Eric Paris <eparis@redhat.com>

This is a new audit feature which only grants processes with
CAP_AUDIT_CONTROL the ability to unset their loginuid.  They cannot
directly set it from a valid uid to another valid uid.  The ability to
unset the loginuid is nice because a priviledged task, like that of
container creation, can unset the loginuid and then priv is not needed
inside the container when a login daemon needs to set the loginuid.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>