[kernel] audit: drop audit_cmd_lock in AUDIT_USER family of cases
Message-id: <9578af9b4c7bcf6ffcffe3a82b9a0654df3a7900.1389385665.git.rgb@redhat.com>
Patchwork-id: 74013
O-Subject: [RHEL7 PATCH 3/4] audit: drop audit_cmd_lock in AUDIT_USER family of cases
Bugzilla: 1019895
RH-Acked-by: Eric Paris <eparis@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

We do not need to hold the audit_cmd_mutex for this family of cases.  The
possible exception to this is the call to audit_filter_user(), so drop the lock
immediately after.  To help in fixing the race we are trying to avoid, make
sure that nothing called by audit_filter_user() calls audit_log_start().  In
particular, watch out for *_audit_rule_match().

This fix will take care of systemd and anything USING audit.  It still means
that we could race with something configuring audit and auditd shutting down.

Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
Reported-by: toshi.okajima@jp.fujitsu.com
Tested-by: toshi.okajima@jp.fujitsu.com
Signed-off-by: Jarod Wilson <jarod@redhat.com>