[block] nvme: Device resume error handling
Message-id: <1389654655-15762-8-git-send-email-dmilburn@redhat.com>
Patchwork-id: 74058
O-Subject: [RHEL7 PATCH 07/13] NVMe: Device resume error handling
Bugzilla: 1049584
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Keith Busch <keith.busch@intel.com>

Adds controller error handling on resume power management. If the device
fails to initialize, the device is queued for a reset. If the reset fails,
a thread is spawned to remove the pci device.

If the device resumes as "busy", the device is responding to admin
commands but will not create IO queues. In this case, we need to remove
the gendisks and free the IO queues since they can't be used and may be
holding bios in their lists.

>From testing, the dma pools require a pci device so this had to change
the pci driver 'remove' to release the dma resources in line with that
call instead of after all references to the device are released.

Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
(cherry picked from commit 9a6b94584de1a0467d85b435df9c744c5c45a270)

BZ 1049584

Signed-off-by: Jarod Wilson <jarod@redhat.com>