[ethernet] sfc: Fix DMA unmapping issue with firmware assisted TSO
Message-id: <1389701203-16561-31-git-send-email-nikolay@redhat.com>
Patchwork-id: 74118
O-Subject: [RHEL7 PATCH BZ 1028038 30/84] sfc: Fix DMA unmapping issue with firmware assisted TSO
Bugzilla: 1028038
RH-Acked-by: Veaceslav Falico <vfalico@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

From: Alexandre Rames <arames@solarflare.com>

When using firmware assisted TSO, we use a single DMA mapping for
the linear area of a TSO skb.

We still have to segment the super-packet and insert a descriptor
containing the original headers before each segment of payload, so we
can unmap the linear area only after the last segment is completed.
The unmapping information for the linear area is therefore associated
with the last header descriptor.

We calculate the DMA address to unmap from using the map length and
the invariant that the end of the DMA mapping matches the end of
the data referenced by the last descriptor.  But this invariant is
broken when there is TCP payload in the linear area.

Fix this by adding and using an explicit dma_offset field.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>
(cherry picked from commit 2acdb92e85412bf791a3c01f2173695b6e0a3f72)
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>