[scsi] isci: fix reset timeout handling
Message-id: <1389818203-22617-1-git-send-email-dmilburn@redhat.com>
Patchwork-id: 74260
O-Subject: [RHEL7 PATCH] isci: fix reset timeout handling
Bugzilla: 1040402
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

This patch has been posted upstream, but not yet pulled in by SCSI
maintainer, Intel has verified fix on RHEL7 test kernel.

http://marc.info/?l=linux-scsi&m=138878043603208&w=2

Remove an erroneous BUG_ON() in the case of a hard reset timeout.  The
reset timeout handler puts the port into the "awaiting link-up" state.
The timeout causes the device to be disconnected and we need to be in
the awaiting link-up state to re-connect the port.  The BUG_ON() made the
incorrect assumption that resets never timeout and we always complete the
reset in the "resetting" state.

Testing this patch also uncovered that libata continues to attempt to
reset the port long after the driver has torn down the context.  Once
the driver has committed to abandoning the link it must indicate to
libata that recovery ends by returning -ENODEV from
->lldd_I_T_nexus_reset().

Cc: <stable@vger.kernel.org>
Cc: Maciej Patelczyk <maciej.patelczyk@intel.com>
Cc: Dave Jiang <dave.jiang@intel.com>
Acked-by: Lukasz Dorau <lukasz.dorau@intel.com>
Reported-by: David Milburn <dmilburn@redhat.com>
Reported-by: Xun Ni <xun.ni@intel.com>
Tested-by: Xun Ni <xun.ni@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

https://bugzilla.redhat.com/show_bug.cgi?id=1040402
https://brewweb.devel.redhat.com/taskinfo?taskID=6807900
Signed-off-by: Jarod Wilson <jarod@redhat.com>