[redhat] Kconfig: really disable RDS this time
Message-id: <7cc3e56e38d194d8350a2323062e1591069e457a.1390407560.git.dledford@redhat.com>
Patchwork-id: 74710
O-Subject: [Patch RHEL7] net/RDS: really disable RDS this time
Bugzilla: 988485
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

The previous attempt to disable RDS only disabled it from the generic
config.  It left it enabled in several of the more specific config
locations.  As a result, it was off on x86_64, but still enabled on
ppc64 and s390x.  Get it everywhere this time.

Resolves: bz988485

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>