[powerpc] Fix VRSAVE handling
Message-id: <20140124200311.31754.84338.sendpatchset@ibm-p7r2-01.lab.bos.redhat.com>
Patchwork-id: 75038
O-Subject: [PATCH RHEL7.0 BZ1042526 3/9] powerpc: Fix VRSAVE handling
Bugzilla: 1052083
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Seth Jennings <sjenning@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=105083

Description:
------------
Since 2002, the kernel has not saved VRSAVE on exception entry and
restored it on exit; rather, VRSAVE gets context-switched in _switch.
This means that when executing in process context in the kernel, the
userspace VRSAVE value is live in the VRSAVE register.

However, the signal code assumes that current->thread.vrsave holds
the current VRSAVE value, which is incorrect.  Therefore, this
commit changes it to use the actual VRSAVE register instead.  (It
still uses current->thread.vrsave as a temporary location to store
it in, as __get_user and __put_user can only transfer to/from a
variable, not an SPR.)

This also modifies the transactional memory code to save and restore
VRSAVE regardless of whether VMX is enabled in the MSR.  This is
because accesses to VRSAVE are not controlled by the MSR.VEC bit,
but can happen at any time.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

RHEL Version Found:
-------------------
7.0

Upstream Status:
----------------
http://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=408a7e08b2112faf687dd629212e42998a7dbe48

----------------
Steve Best
IBM on-site partner

Proposed Patch:
---------------

Signed-off-by: Jarod Wilson <jarod@redhat.com>