[drm] i915: Use the correct GMCH_CTRL register for Sandybridge+
Message-id: <1391119659-12308-44-git-send-email-rclark@redhat.com>
Patchwork-id: 75663
O-Subject: [RHEL7 drm (v2) 43/65] drm/i915: Use the correct GMCH_CTRL register for Sandybridge+
Bugzilla: 1054409
RH-Acked-by: Dave Airlie <airlied@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Chris Wilson <chris@chris-wilson.co.uk>

RH BZ: 1054409
stable: 9a3a1fe

commit a885b3ccc74d8e38074e1c43a47c354c5ea0b01e upstream.

The GMCH_CTRL register (or MGCC in the spec) is at a different address
on Sandybridge, and the address to which we currently write to is
undefined. These stray writes appear to upset (hard hang) my Ivybridge
machine whilst it is in UEFI mode.

Note that the register is still marked as locked RO on Sandybridge, so
vgaarb is still dysfunctional.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Jani Nikula <jani.nikula@linux.intel.com>
Cc: Ville Syrjala <ville.syrjala@linux.intel.com>
Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Jarod Wilson <jarod@redhat.com>